+++ app/(route)/(bottom-nav)/mypage/_components/tab/MyLocationTab/MapInfoBox.tsx
